+++ example/BenchTest/ReactorController.h
@@ -30,7 +30,7 @@ class ReactorController : public chaos::CThreadExecutionTask, private Batch_Cont
     string reactorID;
     boost::shared_ptr<CThread> chaosThread;
     chaos::ui::DeviceController *chaosReactorController;
-    int32_t simulationSpeed;
+    uint64_t simulationSpeed;
     ControllerState state;
     high_resolution_clock::time_point lastExecutionTime;
     high_resolution_clock::time_point currentExecutionTime;
@@ -39,7 +39,7 @@ protected:
     virtual void executeOnThread(const string&) throw(CException);
     
 public:
-    ReactorController(string& _rName, vector<double> *refVec, int refIdx, int32_t _simulationSpeed);
+    ReactorController(string& _rName, vector<double> *refVec, int refIdx, uint64_t _simulationSpeed);
     ~ReactorController();
     void init();
     void deinit();
